Refactor navigation components: move CustomAppBar to common directory and update imports
