+++ src/components/pages/common/index.ts
@@ -1,1 +1,3 @@
+export { default as CustomAppBar } from "./CustomAppBar";
+export { default as Footer } from "./Footer";
 export { default as SocialLinkButton } from "./SocialLinkButton";
